pyproject.toml
[tool.poetry]
package-mode = false
name = "blog"
version = "0.1.0"
description = "Blog API built using Python, Django and DRF."
authors = ["Yousef Abu Shanab <josephyousef249@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
Django = "^5.1.1"
django-cors-headers = "^4.4.0"
django-extensions = "^3.2.3"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
drf-redesign = "^0.4.0"
huggingface-hub = "^0.25.1"
pillow = "^10.4.0"
polars = "^1.8.2"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
